+++ ui-test/tests/Spinner.jsx
@@ -1,16 +1,33 @@
 import React from 'react';
 import {Card, CardTitle, CardContent, Spinner} from '../../index';
 
-const TestSpinner = () => (
-    <Card className="my-own-class" shadow={0} stretch={false} reducedSize>
-        <CardTitle className="my-own-class" border={false} documentLevel="h7">
-            Test Spinner
-        </CardTitle>
-        <CardContent className="my-own-class">
-            <Spinner appearInline />
-            <Spinner appearLocal />
-        </CardContent>
-        <Spinner />
-    </Card>
-);
+class TestSpinner extends React.PureComponent {
+    constructor() {
+        super();
+        this.state = {};
+    }
+
+    render() {
+        return (
+            <Card
+                className="my-own-class"
+                shadow={0}
+                stretch={false}
+                reducedSize>
+                <CardTitle
+                    className="my-own-class"
+                    border={false}
+                    documentLevel="h7">
+                    Test Spinner
+                </CardTitle>
+                <CardContent className="my-own-class">
+                    <Spinner appearInline />
+                    <Spinner appearLocal />
+                </CardContent>
+                <Spinner />
+            </Card>
+        );
+    }
+}
+
 export default TestSpinner;