+++ ui-test/ui-test.jsx
@@ -2,7 +2,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import {Header} from 'react-mdl';
+import {Header, Drawer, Navigation} from 'react-mdl';
+import {BrowserRouter as Router, Route, withRouter} from 'react-router-dom';
+import _ from 'lodash';
 
 // test styles
 import '../style/test.scss';
@@ -13,6 +15,7 @@ import TestAlerts from './tests/Alerts';
 import TestDialogs from './tests/Dialogs';
 import TestIcons from './tests/Icons';
 import TestButtons from './tests/Buttons';
+import TestChips from './tests/Chips';
 import TestInputs from './tests/Inputs';
 import TestSelects from './tests/Selects';
 import TestTabs from './tests/Tabs';
@@ -20,191 +23,165 @@ import TestTable from './tests/Table';
 import TestScrolling from './tests/Scrolling';
 import TestPagination from './tests/Pagination';
 import TestTooltips from './tests/Tooltips';
+import TestCards from './tests/Cards';
+import TestNA from './tests/NA';
+import TestNothing from './tests/Nothing';
+import TestBreadcrumbList from './tests/BreadcrumbList';
+import TestContextMenu from './tests/ContextMenu';
+
 import PerformanceMixin from './../src/mixins/PerformanceMixin';
-import ScrollingMixin from './../src/mixins/ScrollingMixin';
 
 // component
-import {
-    BreadcrumbList,
-    BreadcrumbItem,
-    ContextMenu,
-    MenuItem,
-    Nothing,
-    NotAvailable,
-    Layout,
-    Content,
-    Footer,
-} from '../index';
+import {Layout, Content, Footer} from '../index';
 
 window.enablePerformanceMixingLog = true;
 
 const Page = React.createClass({
     mixins: [PerformanceMixin],
 
-    addContextMenuItem() {
-        this.setState({
-            insertContextMenuItem: true,
-        });
-        console.log('insert MenuItem');
-    },
-
-    removeContextMenuItem() {
-        this.setState({
-            insertContextMenuItem: false,
-        });
-        console.log('remove MenuItem');
-    },
-
-    handleScrollTo(ref) {
-        ScrollingMixin.scrollElementIntoView(this[ref], {topOffset: 100});
-    },
-
     // template rendering
     render() {
         const testcases = [
-            <Nothing key="nothing" />,
-            <TestSpinner
-                key="test-spinner"
-                ref={spinner => {
-                    this.testSpinner = spinner;
-                }}
-            />,
-            <div key="N/A">
-                <p>
-                    This is a test with a block{' '}
-                    <NotAvailable
-                        label="N/A"
-                        description="Not available element"
-                    />{' '}
-                    information.
-                </p>
-                <p>
-                    This is a test with a inline <NotAvailable inline />{' '}
-                    information.
-                </p>
-            </div>,
-            <hr key="spacer1" className="mdl-layout-spacer" />,
-            <TestProgressbar key="progress" />,
-            <hr key="spacer2" className="mdl-layout-spacer" />,
-            <TestAlerts
-                key="alerts"
-                ref={alerts => {
-                    this.testAlerts = alerts;
-                }}
-            />,
-            <hr key="spacer3" className="mdl-layout-spacer" />,
-            <TestScrolling
-                key="scrolling"
-                scrollTestCases={[
-                    {
-                        label: 'Scroll to spinner test',
-                        handleScroll: this.handleScrollTo,
-                        handleRef: 'testSpinner',
-                    },
-                    {
-                        label: 'Scroll to alert test',
-                        handleScroll: this.handleScrollTo,
-                        handleRef: 'testAlerts',
-                    },
-                    {
-                        label: 'Scroll to button test',
-                        handleScroll: this.handleScrollTo,
-                        handleRef: 'testButtons',
-                    },
-                ]}
-            />,
-            <hr key="spacer4" className="mdl-layout-spacer" />,
-            <TestDialogs key="dialogs" />,
-            <hr key="spacer5" className="mdl-layout-spacer" />,
-            <TestTooltips key="tooltips" />,
-            <hr key="spacer6" className="mdl-layout-spacer" />,
-            <TestIcons key="icons" />,
-            <hr key="spacer7" className="mdl-layout-spacer" />,
-            <TestButtons
-                key="buttons"
-                ref={buttons => {
-                    this.testButtons = buttons;
-                }}
-            />,
-            <hr key="spacer8" className="mdl-layout-spacer" />,
-            <TestInputs key="inputs" />,
-            <hr key="spacer8.1" className="mdl-layout-spacer" />,
-            <TestSelects key="selects" />,
-            <hr key="spacer9" className="mdl-layout-spacer" />,
-            <TestTabs key="tabs" />,
-            <hr key="spacer0" className="mdl-layout-spacer" />,
-            <TestPagination key="pagination" />,
-            <hr key="spacerA" className="mdl-layout-spacer" />,
-            <TestTable key="table" />,
-            <hr key="spacerB" className="mdl-layout-spacer" />,
+            {
+                name: 'Context Menu',
+                code: <TestContextMenu key="contextmenu" />,
+            },
+            {
+                name: 'Breadcrumb List',
+                code: <TestBreadcrumbList key="breadcrumblist" />,
+            },
+            {
+                name: 'Nothing',
+                code: <TestNothing key="nothing" />,
+            },
+            {
+                name: 'Spinner',
+                code: <TestSpinner key="test-spinner" />,
+            },
+            {
+                name: 'N/A',
+                code: <TestNA key="N/A" />,
+            },
+            {
+                name: 'Progressbar',
+                code: <TestProgressbar key="progress" />,
+            },
+            {
+                name: 'Alerts',
+                code: <TestAlerts key="alerts" />,
+            },
+            {
+                name: 'Scrolling',
+                code: <TestScrolling key="scrolling" />,
+            },
+            {
+                name: 'Dialogs',
+                code: <TestDialogs key="dialogs" />,
+            },
+            {
+                name: 'Tooltips',
+                code: <TestTooltips key="tooltips" />,
+            },
+            {
+                name: 'Icons',
+                code: <TestIcons key="icons" />,
+            },
+            {
+                name: 'Buttons',
+                code: <TestButtons key="buttons" />,
+            },
+            {
+                name: 'Chips',
+                code: <TestChips key="chips" />,
+            },
+            {
+                name: 'Inputs',
+                code: <TestInputs key="inputs" />,
+            },
+            {
+                name: 'Selects',
+                code: <TestSelects key="selects" />,
+            },
+            {
+                name: 'Tabs',
+                code: <TestTabs key="tabs" />,
+            },
+            {
+                name: 'Pagination',
+                code: <TestPagination key="pagination" />,
+            },
+            {
+                name: 'Table',
+                code: <TestTable key="table" />,
+            },
+            {
+                name: 'Cards',
+                code: <TestCards key="cards" />,
+            },
         ];
 
+        // sort the list of cases for easy searching
+        testcases.sort((a, b) => {
+            if (a.name < b.name) return -1;
+            if (a.name > b.name) return 1;
+            return 0;
+        });
+
+        const path = this.props.location.pathname.substring(1);
+        if (
+            path !== 'all' &&
+            testcases.filter(c => c.name.toLowerCase() === path).length === 0
+        ) {
+            window.location = '/all';
+            return 0;
+        }
+
         return (
-            <Layout fixedHeader>
-                <Header>
-                    <ContextMenu align="left">
-                        <MenuItem className="ownClassName" key="no1">
-                            First First Item
-                        </MenuItem>
-                        <MenuItem>First Second Item</MenuItem>
-                        <MenuItem>First Menu Item 3</MenuItem>
-                        <MenuItem>First Another Menu Item</MenuItem>
-                        <MenuItem onClick={this.addContextMenuItem}>
-                            First Add Another
-                        </MenuItem>
-                        {this.state && this.state.insertContextMenuItem ? (
-                            <MenuItem onClick={this.removeContextMenuItem}>
-                                Remove me
-                            </MenuItem>
-                        ) : (
-                            false
-                        )}
-                    </ContextMenu>
-                    <ContextMenu
-                        align="right"
-                        iconName="add"
-                        tooltip="add property">
-                        <MenuItem>Second First Item</MenuItem>
-                        <MenuItem>Second Second Item</MenuItem>
-                        <MenuItem>Second Menu Item 3</MenuItem>
-                        <MenuItem>Second Another Menu Item</MenuItem>
-                        <MenuItem>Second Alright</MenuItem>
-                    </ContextMenu>
-                    <ContextMenu valign="top">
-                        <MenuItem>Only one menu item</MenuItem>
-                    </ContextMenu>
-                    {this.state && this.state.insertContextMenuItem ? (
-                        <ContextMenu>
-                            <MenuItem>test</MenuItem>
-                        </ContextMenu>
-                    ) : (
-                        false
-                    )}
-                </Header>
-                <Content>
-                    <div className="ecc-application__workspace">
-                        <BreadcrumbList className="my-own-class">
-                            <BreadcrumbItem
-                                onClick={function() {
-                                    alert('Click on breadcrumb item.');
-                                }}>
-                                Button
-                            </BreadcrumbItem>
-                            <BreadcrumbItem href="#">Link</BreadcrumbItem>
-                            <BreadcrumbItem>Span</BreadcrumbItem>
-                        </BreadcrumbList>
-                        <hr className="mdl-layout-spacer" />
-                        {testcases}
-                    </div>
-                </Content>
-                <Footer
-                    company="Eccenca"
-                    version="v0.1.0"
-                    companyUrl="http://eccenca.com"
-                />
-            </Layout>
+            <div>
+                <Layout fixedHeader>
+                    <Header title="Test Cases" />
+                    <Drawer>
+                        <Navigation>
+                            <a href="all">All</a>
+                            {_.map(testcases, ({name}) => (
+                                <a
+                                    href={`${name.toLowerCase()}`}
+                                    key={`${name}`}>
+                                    {name}
+                                </a>
+                            ))}
+                        </Navigation>
+                    </Drawer>
+                    <Content>
+                        {_.map(testcases, ({name, code}) => (
+                            <Route
+                                path={`/(all|${name.toLowerCase()})`}
+                                key={`${name}`}
+                                render={() => (
+                                    <div className="uitest-divmargin">
+                                        {code}
+                                    </div>
+                                )}
+                            />
+                        ))}
+                    </Content>
+                    <Footer
+                        company="Eccenca"
+                        version="v0.1.0"
+                        companyUrl="http://eccenca.com"
+                    />
+                </Layout>
+            </div>
         );
     },
 });
 
-ReactDOM.render(<Page />, document.getElementById('react'));
+const PageWithRouter = withRouter(Page);
+
+const PageWithRouting = props => (
+    <Router>
+        <PageWithRouter {...props} />
+    </Router>
+);
+
+ReactDOM.render(<PageWithRouting />, document.getElementById('react'));