+++ CHANGELOG.md
@@ -4,8 +4,16 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ## [Unreleased]
 
+### Added
+
+- Routing to test cases
+- New Cards test
+
 ### Changed
-- cut value description in `<AutoCompleteBox/>` on the left side if value is to big to get shown completely
+
+- Cut value description in `<AutoCompleteBox/>` on the left side if value is to big to get shown completely
+- Redesign the content using MDL sidebar with `<Drawer>` and `<Navigation>` elements
+- Separate chips test from buttons test
 
 ## [3.3.1] 2018-12-11